+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed path not defaulting to `/` when using custom base url
+- Fixed navigation happening when left-clicking on a link while Shift is pressed
 
 
 ## [6.0.8] - 2026-04-18